refactor: modularize UI into separate module
