+++ src/ui/mod.rs
@@ -1,0 +1,4 @@
+mod app;
+pub mod editor;
+
+pub use app::{App, run};
